versions/v1_frozen/requirements.txt
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
PyYAML==6.0.2
scikit-learn==1.5.1
